+++ packages/graph-engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tokens-studio/graph-engine
 
+## 0.10.0
+
+### Minor Changes
+
+- eefa966: add ungroup node, add select single token node, fix input issue on group and extract tokens
+- eaf05cd: Add Contrasting from Set node to return the first element of an array that has sufficient contrast
+
 ## 0.9.0
 
 ### Minor Changes